+++ decidim-conferences/app/cells/decidim/conferences/join_conference_button_cell.rb
@@ -34,7 +34,7 @@ module Decidim
       end
 
       def i18n_join_text
-        return I18n.t("join", scope: "decidim.conferences.conference.show") if model.registrations_enabled?
+        return I18n.t("join", scope: "decidim.conferences.conference.show") if model.has_available_slots?
         I18n.t("no_slots_available", scope: "decidim.conferences.conference.show")
       end
     end