+++ decidim-conferences/lib/decidim/conferences/participatory_space.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 Decidim.register_participatory_space(:conferences) do |participatory_space|
-  participatory_space.icon = "decidim/conferences/icon.svg"
+  participatory_space.icon = "decidim/conferences/conference.svg"
   participatory_space.model_class_name = "Decidim::Conference"
 
   participatory_space.participatory_spaces do |organization|