+++ decidim-conferences/spec/events/decidim/conferences/update_conference_event_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+describe Decidim::Conferences::UpdateConferenceEvent do
+  let(:resource) { create :conference }
+  let(:event_name) { "decidim.events.conferences.conference_updated" }
+
+  include_context "when a simple event"
+  it_behaves_like "a simple event"
+end